- moved getBlogPosts to WikiAPI
