+++ lib/groups/groupsAPI.js
@@ -8,7 +8,6 @@ var async = require('async');
 var validation = conf.get('beans').get('validation');
 var groupstore = conf.get('beans').get('groupstore');
 var misc = conf.get('beans').get('misc');
-var Git = conf.get('beans').get('gitmech');
 
 var sympaClient;
 //Just checking if remote has been configured
@@ -198,44 +197,5 @@ module.exports = {
 
   isReserved: function (groupname) {
     return new RegExp('^edit$|^new$|^checkgroupname$|^submit$|^administration$|[^\\w-]', 'i').test(groupname);
-  },
-
-  getBlogposts : function (groupname, callback) {
-    Git.lsblogposts(groupname, function (err, result) {
-      if (err) { return callback(err); }
-
-      if (result.length === 0) {
-        return callback(err, []);
-      }
-
-      async.map(result,
-        function (path, mapCallback) {
-          Git.readFile(path, 'HEAD', function (err, rawPost) {
-            if (err) { return mapCallback(err); }
-
-            var parsedPost = misc.parseBlogPost(rawPost);
-            if (parsedPost !== undefined) {
-              parsedPost.path = path.substring(0, path.length - 3);
-            }
-            mapCallback(null, parsedPost);
-          });
-        },
-        function (err, unsortedPosts) {
-          if (err) { return callback(err); }
-
-          var postsWithoutUndefined = _.filter(unsortedPosts, function (post) {
-            return post !== undefined;
-          });
-
-          var postsSortedByDate = _.sortBy(postsWithoutUndefined, function (post) {
-            if (!post.date.isValid()) {
-              return Number.MAX_VALUE;
-            }
-            // _.sortBy is always in ascending order and we want the latest post first, so use negation
-            return -post.date.unix();
-          });
-          callback(null, postsSortedByDate);
-        });
-    });
   }
 };
