+++ lib/groups/index.js
@@ -5,6 +5,7 @@ var beans = require('nconf').get('beans');
 var async = require('async');
 
 var groupsAPI = beans.get('groupsAPI');
+var wikiAPI = beans.get('wikiAPI');
 var Group = beans.get('group');
 var groupsAndMembers = beans.get('groupsAndMembersAPI');
 var statusmessage = beans.get('statusmessage');
@@ -93,7 +94,7 @@ module.exports = function (app) {
 
     groupsAndMembers.getGroupAndMembersForList(req.params.groupname, function (err, group) {
       if (err || !group) { return next(err); }
-      groupsAPI.getBlogposts(req.params.groupname, function (err, blogposts) {
+      wikiAPI.getBlogpostsForGroup(req.params.groupname, function (err, blogposts) {
         res.render('get', {group: group, users: group.members,
           userIsGroupMember: groupsAndMembers.userIsInMemberList(registeredUserId(req), group.members),
           organizers: group.organizers,
