+++ lib/wiki/wikiAPI.js
@@ -4,6 +4,9 @@ var Diff = require('./gitDiff');
 var Path = require('path');
 var beans = require('nconf').get('beans');
 var Git = beans.get('gitmech');
+var async = require('async');
+var misc = beans.get('misc');
+var _ = require("underscore");
 
 module.exports = {
 
@@ -87,6 +90,45 @@ module.exports = {
         }
       });
       callback(null, result);
+    });
+  },
+
+  getBlogpostsForGroup : function (groupname, callback) {
+    Git.lsblogposts(groupname, function (err, result) {
+      if (err) { return callback(err); }
+
+      if (result.length === 0) {
+        return callback(err, []);
+      }
+
+      async.map(result,
+        function (path, mapCallback) {
+          Git.readFile(path, 'HEAD', function (err, rawPost) {
+            if (err) { return mapCallback(err); }
+
+            var parsedPost = misc.parseBlogPost(rawPost);
+            if (parsedPost !== undefined) {
+              parsedPost.path = path.substring(0, path.length - 3);
+            }
+            mapCallback(null, parsedPost);
+          });
+        },
+        function (err, unsortedPosts) {
+          if (err) { return callback(err); }
+
+          var postsWithoutUndefined = _.filter(unsortedPosts, function (post) {
+            return post !== undefined;
+          });
+
+          var postsSortedByDate = _.sortBy(postsWithoutUndefined, function (post) {
+            if (!post.date.isValid()) {
+              return Number.MAX_VALUE;
+            }
+            // _.sortBy is always in ascending order and we want the latest post first, so use negation
+            return -post.date.unix();
+          });
+          callback(null, postsSortedByDate);
+        });
     });
   }
 };
